+++ .gitignore
@@ -1,0 +1,2 @@
+/gh-epr
+/gh-epr.exe
