+++ go.mod
@@ -1,3 +1,3 @@
-module gh-epr
+module github.com/wannabehero/gh-epr
 
 go 1.24.2
